refactor: expose blockHub in EditorContainer
