+++ packages/blocks/src/index.ts
@@ -31,6 +31,7 @@ export * from './__internal__/utils/std.js';
 export * from './__internal__/utils/query.js';
 export * from './__internal__/utils/lit.js';
 export * from './__internal__/rich-text/rich-text-operations.js';
+export * from './components/index.js';
 export type { BaseService } from './__internal__/service.js';
 export { getServiceOrRegister } from './__internal__/service.js';
