+++ packages/blocks/src/page-block/default/default-page-block.ts
@@ -40,8 +40,6 @@ import { assertExists } from '@blocksuite/global/utils';
 import type { DragHandle } from '../../components/index.js';
 import { BLOCK_ID_ATTR } from '@blocksuite/global/config';
 import { bindHotkeys, removeHotkeys } from '../utils/bind-hotkey.js';
-import type { BlockHub } from '../../components/index.js';
-import { createBlockHub } from '../utils/components.js';
 
 export interface EmbedEditingState {
   position: { x: number; y: number };
@@ -148,10 +146,8 @@ export class DefaultPageBlockComponent
    */
   components: {
     dragHandle: DragHandle | null;
-    blockHub: BlockHub | null;
   } = {
     dragHandle: null,
-    blockHub: null,
   };
 
   @property()
@@ -363,43 +359,6 @@ export class DefaultPageBlockComponent
     );
   };
 
-  private _initBlockHub = () => {
-    if (
-      this.page.awarenessStore.getFlag('enable_block_hub') &&
-      !this.components.blockHub
-    ) {
-      this.components.blockHub = createBlockHub(
-        this,
-        this.signals.updateSelectedRects
-      );
-      this.components.blockHub.getAllowedBlocks = () =>
-        getAllowSelectedBlocks(this.model);
-    }
-    this._disposables.add(
-      this.page.awarenessStore.signals.update.subscribe(
-        msg => msg.state?.flags.enable_block_hub,
-        enable => {
-          if (enable) {
-            if (!this.components.blockHub) {
-              this.components.blockHub = createBlockHub(
-                this,
-                this.signals.updateSelectedRects
-              );
-              this.components.blockHub.getAllowedBlocks = () =>
-                getAllowSelectedBlocks(this.model);
-            }
-          } else {
-            this.components.blockHub?.remove();
-            this.components.blockHub = null;
-          }
-        },
-        {
-          filter: msg => msg.id === this.page.doc.clientID,
-        }
-      )
-    );
-  };
-
   private _getViewportScrollOffset() {
     const container = this.defaultViewportElement;
     return {
@@ -471,14 +430,13 @@ export class DefaultPageBlockComponent
   override connectedCallback() {
     super.connectedCallback();
     this._initDragHandle();
-    this._initBlockHub();
+    // this._initBlockHub();
   }
 
   override disconnectedCallback() {
     super.disconnectedCallback();
     this._disposables.dispose();
     this.components.dragHandle?.remove();
-    this.components.blockHub?.remove();
 
     removeHotkeys();
     this.selection.dispose();
