+++ packages/blocks/src/page-block/edgeless/edgeless-page-block.ts
@@ -35,9 +35,6 @@ import { styleMap } from 'lit/directives/style-map.js';
 import type { SurfaceBlockModel } from '../../surface-block/surface-model.js';
 import { SurfaceManager } from '@blocksuite/phasor';
 import { BLOCK_ID_ATTR, HOTKEYS } from '@blocksuite/global/config';
-import type { BlockHub } from '../../components/index.js';
-import { getAllowSelectedBlocks } from '../default/utils.js';
-import { createBlockHub } from '../utils/components.js';
 
 export interface EdgelessContainer extends HTMLElement {
   readonly page: Page;
@@ -123,12 +120,6 @@ export class EdgelessPageBlockComponent
 
   getService = getService;
 
-  components: {
-    blockHub: BlockHub | null;
-  } = {
-    blockHub: null,
-  };
-
   private _disposables = new DisposableGroup();
   private _selection!: EdgelessSelectionManager;
 
@@ -194,37 +185,6 @@ export class EdgelessPageBlockComponent
     this._syncSurfaceViewport();
   }
 
-  private _initBlockHub = () => {
-    if (
-      this.page.awarenessStore.getFlag('enable_block_hub') &&
-      !this.components.blockHub
-    ) {
-      this.components.blockHub = createBlockHub(this);
-      this.components.blockHub.getAllowedBlocks = () =>
-        getAllowSelectedBlocks(this.pageModel);
-    }
-    this._disposables.add(
-      this.page.awarenessStore.signals.update.subscribe(
-        msg => msg.state?.flags.enable_block_hub,
-        enable => {
-          if (enable) {
-            if (!this.components.blockHub) {
-              this.components.blockHub = createBlockHub(this);
-              this.components.blockHub.getAllowedBlocks = () =>
-                getAllowSelectedBlocks(this.pageModel);
-            }
-          } else {
-            this.components.blockHub?.remove();
-            this.components.blockHub = null;
-          }
-        },
-        {
-          filter: msg => msg.id === this.page.doc.clientID,
-        }
-      )
-    );
-  };
-
   update(changedProperties: Map<string, unknown>) {
     if (changedProperties.has('mouseRoot') && changedProperties.has('page')) {
       this._selection = new EdgelessSelectionManager(this);
@@ -257,7 +217,6 @@ export class EdgelessPageBlockComponent
       this.requestUpdate();
     });
     this._disposables.add(historyDisposable);
-    this._initBlockHub();
     this._bindHotkeys();
 
     tryUpdateFrameSize(this.page, this.viewport.zoom);
@@ -279,7 +238,6 @@ export class EdgelessPageBlockComponent
 
   disconnectedCallback() {
     super.disconnectedCallback();
-    this.components.blockHub?.remove();
 
     this.signals.updateSelection.dispose();
     this.signals.viewportUpdated.dispose();
